more tests and raise exception for subsettings of None
